+++ local-s3-rest/src/main/java/com/robothy/s3/rest/handler/DeletePublicAccessBlockController.java
@@ -7,6 +7,7 @@ import com.robothy.s3.core.service.BucketService;
 import com.robothy.s3.rest.assertions.RequestAssertions;
 import com.robothy.s3.rest.service.ServiceFactory;
 import com.robothy.s3.rest.utils.ResponseUtils;
+import io.netty.handler.codec.http.HttpResponseStatus;
 
 /**
  * Handle <a href="https://docs.aws.amazon.com/AmazonS3/latest/API/API_DeletePublicAccessBlock.html">DeletePublicAccessBlock</a>.
@@ -24,6 +25,7 @@ class DeletePublicAccessBlockController implements HttpRequestHandler {
   public void handle(HttpRequest request, HttpResponse response) throws Exception {
     String bucketName = RequestAssertions.assertBucketNameProvided(request);
     bucketService.deletePublicAccessBlock(bucketName);
+    response.status(HttpResponseStatus.NO_CONTENT);
 
     ResponseUtils.addDateHeader(response);
     ResponseUtils.addServerHeader(response);